+++ tests/test_serializers__exclude_fields_mixin.py
@@ -172,3 +172,23 @@ class ExcludeFieldsSerializerMixinTests(SerializerMixinTestCase):
     def test_missing_child_key_many(self):
         with self.assertRaises(ValueError):
             self.serialize(exclude={'skus__not_found'})
+
+    def test_field_ordering_unchanged_root(self):
+        root_1 = self.serialize(exclude=('id', 'manufacturer'))
+        root_2 = self.serialize(exclude=('manufacturer', 'id'))
+
+        # The other of the fields is the same despite the `exclude` ordering
+        self.assertEqual(root_1.keys(), root_2.keys())
+
+        # And that order matches the serializer field ordering
+        self.assertListEqual(root_1.keys(), ['name', 'skus'])
+
+    def test_field_ordering_unchanged_nested(self):
+        child_1 = self.serialize(exclude=('skus__variant',))
+        child_2 = self.serialize(exclude=('skus__owners',))
+
+        keys_1 = child_1['skus'][0].keys()
+        self.assertListEqual(keys_1, ['id', 'owners'])
+
+        keys_2 = child_2['skus'][0].keys()
+        self.assertListEqual(keys_2, ['id', 'variant'])