+++ tests/test_serializers__only_fields_mixin.py
@@ -230,3 +230,23 @@ class OnlyFieldsSerializerMixinTests(SerializerMixinTestCase):
     def test_error_serialize_all_and_specific(self):
         with self.assertRaises(ValueError):
             self.serialize(only={'manufacturer', 'manufacturer__name'})
+
+    def test_field_ordering_unchanged_root(self):
+        root_1 = self.serialize(only=('name', 'id', 'manufacturer'))
+        root_2 = self.serialize(only=('manufacturer', 'id', 'name'))
+
+        # The other of the fields is the same despite the `only` ordering
+        self.assertEqual(root_1.keys(), root_2.keys())
+
+        # And that order matches the serializer field ordering
+        self.assertListEqual(root_1.keys(), ['id', 'name', 'manufacturer'])
+
+    def test_field_ordering_unchanged_nested(self):
+        child_1 = self.serialize(only=('skus__variant', 'skus__owners'))
+        child_2 = self.serialize(only=('skus__owners', 'skus__variant'))
+
+        keys_1 = child_1['skus'][0].keys()
+        keys_2 = child_2['skus'][0].keys()
+
+        self.assertEqual(keys_1, keys_2)
+        self.assertListEqual(keys_1, ['variant', 'owners'])